src/requirements.txt
aniso8601==7.0.0
antlr4-python3-runtime==4.7.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
geojson==2.5.0
graphene==2.1.9
graphene-sqlalchemy==2.1.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
promise==2.3
pydantic~=1.10.13
PyYAML==6.0.1
Rx==1.6.1
singledispatch~=3.7.0
git+https://github.com/Amsterdam/flask-audit-log.git@v0.1.0a-rc1#egg=datapunt-flask-audit-log
git+https://github.com/Amsterdam/GOB-Core.git@v2.26.1
